-Fix compile warnings. -call setAlphaF instead of setAlpha if we are going to use qreal.
svn path=/trunk/KDE/kdebase/workspace/; revision=755226

+++ plasma/applets/kickoff/ui/itemdelegate.cpp
@@ -91,8 +91,11 @@ void ItemDelegate::paint(QPainter *painter, const QStyleOptionViewItem& option, 
         QFontMetrics titleMetrics(titleFont);
         QFontMetrics subTitleMetrics(subTitleFont);
         QRect textAreaRect = contentRect;
-        qreal actualTextWidth = qMax(titleMetrics.width(titleText), subTitleMetrics.width(subTitleText));
-        textAreaRect.adjust(decorationRect.width() + ICON_TEXT_MARGIN - 3, 0, -(titleRect.width() - actualTextWidth) + 3, 1);
+        int actualTextWidth = qMax(titleMetrics.width(titleText), subTitleMetrics.width(subTitleText));
+        textAreaRect.adjust(decorationRect.width() + ICON_TEXT_MARGIN - 3,
+                            0,
+                            -(titleRect.width() - actualTextWidth) + 3,
+                            1);
         // use a slightly translucent version of the palette's highlight color
         // for the background
         backgroundColor.setAlphaF(0.5);
